feat: NetworkModel 생성
